Finish day 11 part 2 but break part 1 :/

+++ AdventOfCode2022Solutions/Day11/Operations/Sum.cs
@@ -6,7 +6,7 @@ namespace AdventOfCode2022Tests.Day11
     {
         public Sum(int number) : base(number) { }
 
-        public override long Execute(long itemValue)
+        public override int Execute(int itemValue)
         {
             return Number + itemValue;
         }
